Emitter Refactoring
Refactored Code so that rather than GameManager to PlayerData to Field,
the flow now goes GameManager to PlayerData to GameManager to Field.
Essentially, PlayerData is now responsible just for emitting what the
result of a roll is, GameManager is then responsible for checking for
Strategy dependent on this result, and then calling field to process the
eventual result.

Additionally, fixed an error where homeruns didn't work when the value
of the roll was higher than the low value of homerun.