Implemented Strategy Card #40 and #12
Implemented Card #40 as a proof of concept that the changes made
yesterday actually work, fixed an error where new lines weren't being
read correctly in StrategyCard and also getRandomCard does actually get
a random card now.

Implemented Card #12 for the purposes of implementing a way of
implementing or conditions. The way I did it is sloppy and I'm
considering changing it as well as disposing of the parse postcondition
throwing exceptions, instead it seems like I can just return false
somehow. I don't actually know for sure that Card #12 works yet because
of the exceptions being thrown.

As my number of cards increases I need to start programming actual
mechanics for handling the cards, maybe the start of AI.