How I Learned to Stop Worrying and Use External Swing Libraries
Added SwingX to the class path because I've decided Swing is a pain in
the neck

Implemented Comparable in PlayerData.

Created a PlayerData Comparator class. This allows sorting of cards
based on however I decided to compare cards.

Modified DraftManager so that drafted players are not actually removed
from the class, but are instead thrown in a hidden bin. Included a
method that allows a reset of a draft pool. getPool also now returns a
seperate copy of the draft pool because the usage I seem to be using it
for I only ever want a copy.

Reorderd methods in MainRunner for readability, disposed of all JLists
and changed them to JXLists, changed all code needed to allow continued
workability. The list of players in the UI now appears in order of set
number.