Can move players around in the lineup editor now
I cheated and used a try catch for verifying integer, I should go back
and change that in the future.

Also import export is now configured for lineup and position.

I don't think I can go any further until I rewreite the entire codebase
so that lineup is of playerdata and not hitterdata, else I think there
will be problems not only with my lineup editor, but also implementing
some cards, in particular the ones that say "When a pitcher is batting"
:+1:

Emojis in github, yikes.