Glitch Fixes and Recording
Fixed the missing actual effect on 15. Amusingly, the way the functional
test is set up, if 15 is selected the test goes infinitely long since
any out will just be rerolled.

Fixed use strategy mistaking which team is on offense.

Card 13 is causing problems by being used as a defensive card after an
offensive card is used to proc a reroll. I think restricting offensive
cards to only be used on offense and defensive cards to only be used on
defense, this issue and similar other ones should go away naturally.