Major Update
Created Enummeration for Positions and removed other methods that
translated positions.

Finished implementation of basic functionality in the lineup editor in
the main GUI. With that framework laid, it's now time for the main
gameplay part of the GUI to take shape, although I also need some text
fields for file import/export.

Broke insertPlayer in lineup manager into 2 separate methods, 1 for
field and 1 for lineup. Will need to eventually put in condition to
enforce rule consistency but in the meantime this is fine.

Implemented methods in lineup manager to check if a given player is in
the lineup or field, returning the index that the player is found at.

Refactored some methods in mainRunner.

Added several getter methods to StrategyCard.

Updated comments in Range.

Updated functional test using new lineupmanager methods.

Created new, fancy comment checklist.

Removed dead imports.